lab7/requirements.txt
prettytable-3.12.0
psycopg2-binary-2.9.10